Add files via upload
This is what I have so far. I have also created a clone of the Vivino API GitHub repository. 

Somewhat of a coders block if you will, I didn't have much trouble implementing the stock APIs I used, but for some reason I'm sort of unable to figure out the solution for this. Thanks for taking a look.